+++ src/main/java/com/sammy/malum/common/blocks/itemstand/ItemStandTileEntity.java
@@ -6,6 +6,7 @@ import com.sammy.malum.core.systems.tileentities.SimpleInventory;
 import com.sammy.malum.core.systems.tileentities.SimpleInventoryTileEntity;
 import com.sammy.malum.core.systems.tileentities.SimpleTileEntity;
 import net.minecraft.util.Direction;
+import net.minecraft.util.math.vector.Vector3d;
 import net.minecraft.util.math.vector.Vector3f;
 
 import static net.minecraft.state.properties.BlockStateProperties.FACING;
@@ -27,10 +28,10 @@ public class ItemStandTileEntity extends SimpleInventoryTileEntity
         };
     }
     
-    public static Vector3f itemOffset(SimpleTileEntity tileEntity)
+    public static Vector3d itemOffset(SimpleTileEntity tileEntity)
     {
         Direction direction = tileEntity.getBlockState().get(FACING);
-        Vector3f directionVector = new Vector3f(direction.getXOffset(), 0.5f, direction.getZOffset());
-        return new Vector3f(0.5f - directionVector.getX() * 0.25f, directionVector.getY(), 0.5f - directionVector.getZ() * 0.25f);
+        Vector3d directionVector = new Vector3d(direction.getXOffset(), 0.5f, direction.getZOffset());
+        return new Vector3d(0.5f - directionVector.getX() * 0.25f, directionVector.getY(), 0.5f - directionVector.getZ() * 0.25f);
     }
 }
